Improve collision detection logic for vendor prefixes.
Before, the module was combining the vendor prefix and property
definition into a single identifier internally, and comparing
the two. Now, we strip the vendor prefix from the internal
checking, so that the module will find a conflict between a
vendor prefixed property and a specification property.

For example, it used to merge properties that contained
a `background` definition along with `-webkit-background-clip`.

Fixes #24.